+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - added `EventEmitter` as base class for `Interval`
 - added `AlarmClock` for defined alarm handling (once a day), with optional raising an event
 - ntp module fires `ntp` now also after timezone change
+- added `Button` for button optional with sending event. supports debouncing
 -